# apps/ebpco-api/contract/citizen-endpoints.openapi.yaml
...
openapi: 3.1.0
info:
  title: eBPCO citizen endpoints — permit, documents, resubmission
  version: 1.0.0
  description: |
    Shipped in eBPCOBackend: C-1 at f7eb40e, C-2 at 5a0f18a, D-8 at 18028a8.

    All three require a citizen bearer token. All three answer 404 for an
    application that is not the caller's, which is the SAME 404 as one that does
    not exist: a reference number that resolves differently would confirm that a
    neighbour has applied.
paths:
...
  /applications/{applicationId}/documents/{documentId}/resubmit:
    post:
      operationId: resubmitDocument
      summary: Replace one document with a newly supplied file
      description: |
        APPENDS, never overwrites: the replacement is a new document pointing at
        what it replaces, and the old one keeps its rejection and reason.

        `Idempotency-Key` is REQUIRED and must be a UUID. A retry with the same
        key replays the identical 201 body rather than creating a second
        document. The FILE is part of the fingerprint, so the same key carrying a
        different replacement is refused 409 — never reuse a key across files.

        Size: `contentBase64` is capped at 40,000,000 characters by this schema,
        but the server's body limit is 1MB by default and base64 inflates by
        about a third, so the real ceiling on a FILE is roughly 750KB. Over it,
        the HTTP adapter answers a bare 413 that is NOT a problem document.
      security: [{ bearerAuth: [documents:write] }]
      parameters:
        - { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [fileName, label, contentBase64]
              properties:
                fileName: { type: string, minLength: 1, maxLength: 255 }
                label: { type: string, minLength: 1, maxLength: 200 }
                contentBase64: { type: string, minLength: 1, maxLength: 40000000 }
      responses:
        '201':
          description: The replacement was stored and linked.
          content: { application/json: { schema: { $ref: '#/components/schemas/ResubmitResult' } } }
        '404':
          description: Not the caller's application, or no such document ON that application.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
        '409':
          description: |
            Already replaced, already accepted by an officer, or the
            Idempotency-Key was used for a different request. `detail` says which.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
        '422':
          description: The file failed a malware check and was not stored.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ResubmitResult:
      type: object
      additionalProperties: false
      required: [documentId, supersedesDocumentId, status, removedMetadata]
      properties:
        documentId: { type: string, format: uuid, description: The NEW document. }
        supersedesDocumentId: { type: string, format: uuid }
        status: { type: string, enum: ['Pending', 'Approved', 'Rejected', 'Missing'] }
        removedMetadata:
          type: array
          items: { type: string }
          description: |
            Metadata stripped from the file — EXIF, GPS. Returned rather than
            silently dropped: a photograph of a site carries its coordinates, and
            an applicant is entitled to know the LGU removed them.
    Problem:
      type: object
      description: RFC 9457. `detail` is the human-readable half.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }